pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentguard"
version = "1.0.0"
description = "AI Agent Behavior Compliance Guardian - 轻量级AI Agent行为准则管理与合规检查工具"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "AgentGuard Team", email = "agentguard@example.com"}
]
keywords = ["ai", "agent", "compliance", "guardrails", "llm", "safety"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.9"
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "jsonschema>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
agentguard = "agentguard.cli:main"
agd = "agentguard.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/agentguard"
Documentation = "https://github.com/yourusername/agentguard#readme"
Repository = "https://github.com/yourusername/agentguard"
Issues = "https://github.com/yourusername/agentguard/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentguard*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "D"]
ignore = ["D100", "D104"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=agentguard --cov-report=term-missing"
